Major changes in LastLayer tool. Coefficients called in xml. Iteration now done in the code. Also a status code as follows:
eLeak=-1 not enough energy in the CAL available to use this method.

eLeak=-2 there is no tracker information. We will later use the CAL information
         in this case and give it a special flag.

eLeak=-3 the reconstructed angle is not in the valid range.

eLeak=-4 the impact parameter on the CAL is such that there is a need for crack
         correction.